chore: cleanup

+++ test/c/array.c
@@ -1,5 +1,5 @@
 int main () {
-  int* foo = malloc(sizeof(int) * 5);
+  double * foo = malloc(sizeof(double) * 5);
   foo[0] = 1;
   foo[3] = 6;
 }
